pyproject.toml
[project]
name = "voice-learning-backend"
version = "0.1.0"
description = "Backend for Voice-to-Voice Learning Platform"
requires-python = "==3.11.*"
dependencies = [
    "fastapi[standard]==0.115.11",
    "uvicorn[standard]==0.34.0",
    "python-multipart==0.0.20",
    "httpx==0.28.1",
    "openai==1.59.9",
    "faiss-cpu==1.9.0",
    "numpy==2.2.3",
    "pdfplumber==0.11.5",
    "pytesseract==0.3.13",
    "pillow==11.1.0",
    "python-dotenv==1.0.1",
    "loguru==0.7.3",
]